Updated to latest version
Almost at the real release.
All that's left to do is choose a router, add a few new namespaces, and
do a final restructuring.

+++ _plugins/testPlugin/quickUninstall.php
@@ -3,18 +3,18 @@ if(!$test){
     unset($_SESSION['testRandomNumber']);
     unset($_SESSION['testSetting']);
 
+    if(!$local){
+        //Create a PDO connection
+        $sqlSettings = new IOFrame\settingsHandler($this->settings->getSetting('absPathToRoot').SETTINGS_DIR_FROM_ROOT.'/sqlSettings/');
+        $conn = IOFrame\prepareCon($sqlSettings);
+        // set the PDO error mode to exception
+        $conn->setAttribute(PDO::ATTR_ERRMODE, PDO::ERRMODE_EXCEPTION);
 
-    //Create a PDO connection
-    $sqlSettings = new IOFrame\settingsHandler($this->settings->getSetting('absPathToRoot').SETTINGS_DIR_FROM_ROOT.'/sqlSettings/');
-    $conn = IOFrame\prepareCon($sqlSettings);
-    // set the PDO error mode to exception
-    $conn->setAttribute(PDO::ATTR_ERRMODE, PDO::ERRMODE_EXCEPTION);
-
-    // INITIALIZE CORE VALUES TABLE
-    /* Literally just the pivot time for now. */
-    $makeTB = $conn->prepare("DROP TABLE IF EXISTS ".$this->sqlHandler->getSQLPrefix()."TEST_TABLE CASCADE;");
-    $makeTB->execute();
-
+        // INITIALIZE CORE VALUES TABLE
+        /* Literally just the pivot time for now. */
+        $makeTB = $conn->prepare("DROP TABLE IF EXISTS ".$this->sqlHandler->getSQLPrefix()."TEST_TABLE CASCADE;");
+        $makeTB->execute();
+    }
 }
 else{
     echo 'quickUninstall activates here!'.EOL;
